🚚 Move player type to own folder
